+++ src/javascript/9000-main.js
@@ -1,5 +1,6 @@
-s.version.set(0,4,9);
+s.version.set(0,4,11);
 s.savesMgr.setBlock(true); // block loading of saves from old versions
+s.savesMgr.setBkMarks(8); // allow 8 bookmarks instead of 1
 st.page.setContinuous(true); // make embedding passages the default
 
 // build menu